+++ public/globalvars_base.h
@@ -1,6 +1,6 @@
 //========= Copyright Valve Corporation, All rights reserved. ============//
 //
-// Purpose: 
+// Purpose:
 //
 //=============================================================================//
 
@@ -21,7 +21,7 @@ class CGlobalVarsBase
 public:
 
 	CGlobalVarsBase( bool bIsClient );
-	
+
 	// This can be used to filter debug output or to catch the client or server in the act.
 	bool IsClient() const;
 
@@ -29,8 +29,8 @@ public:
 	int GetNetworkBase( int nTick, int nEntity );
 
 public:
-	
-	// Absolute time (per frame still - Use Plat_FloatTime() for a high precision real time 
+
+	// Absolute time (per frame still - Use Plat_FloatTime() for a high precision real time
 	//  perf clock, but not that it doesn't obey host_timescale/host_framerate)
 	float			realtime;
 	// Absolute frame counter
@@ -38,23 +38,23 @@ public:
 	// Non-paused frametime
 	float			absoluteframetime;
 
-	// Current time 
+	// Current time
 	//
 	// On the client, this (along with tickcount) takes a different meaning based on what
 	// piece of code you're in:
-	// 
+	//
 	//   - While receiving network packets (like in PreDataUpdate/PostDataUpdate and proxies),
 	//     this is set to the SERVER TICKCOUNT for that packet. There is no interval between
 	//     the server ticks.
 	//     [server_current_Tick * tick_interval]
 	//
-	//   - While rendering, this is the exact client clock 
+	//   - While rendering, this is the exact client clock
 	//     [client_current_tick * tick_interval + interpolation_amount]
 	//
 	//   - During prediction, this is based on the client's current tick:
 	//     [client_current_tick * tick_interval]
-	float			curtime;
-	
+	double			curtime;
+
 	// Time spent on last server or client frame (has nothing to do with think intervals)
 	float			frametime;
 	// current maxplayers setting
@@ -64,7 +64,7 @@ public:
 	int				tickcount;
 
 	// Simulation tick interval
-	float			interval_per_tick;
+	double			interval_per_tick;
 
 	// interpolation amount ( client-only ) based on fraction of next tick which has elapsed
 	float			interpolation_amount;
@@ -80,11 +80,11 @@ private:
 	bool			m_bClient;
 
 	// 100 (i.e., tickcount is rounded down to this base and then the "delta" from this base is networked
-	int				nTimestampNetworkingBase;   
+	int				nTimestampNetworkingBase;
 	// 32 (entindex() % nTimestampRandomizeWindow ) is subtracted from gpGlobals->tickcount to set the networking basis, prevents
 	//  all of the entities from forcing a new PackedEntity on the same tick (i.e., prevents them from getting lockstepped on this)
-	int				nTimestampRandomizeWindow;  
-	
+	int				nTimestampRandomizeWindow;
+
 };
 
 inline int CGlobalVarsBase::GetNetworkBase( int nTick, int nEntity )
